requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
distro==1.9.0
dnspython==1.16.0
eventlet==0.31.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
gevent==25.4.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
mss==10.0.0
numpy==2.2.5
openai==1.77.0
pandas==2.2.3
pandas-stubs==2.2.3.250308
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-socketio==5.13.0
pytz==2025.2
requests==2.32.3
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
tqdm==4.67.1
types-pytz==2025.2.0.20250326
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
